molecular-ai/src/train.py
"""
Main training script for molecular generation models.
"""

import os
import logging
from typing import Dict, Any
from google.cloud import aiplatform
from google.cloud.aiplatform import hyperparameter_tuning as hpt

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

def train_model(
    project_id: str,
    location: str,
    dataset_path: str,
    model_type: str = "vae",
    hyperparameters: Dict[str, Any] = None,
) -> None:
    """
    Train a molecular generation model using Vertex AI.
    
    Args:
        project_id: GCP project ID
        location: GCP region
        dataset_path: Path to training data
        model_type: Type of model to train (vae, gan, gnn)
        hyperparameters: Model hyperparameters
    """
    # Initialize Vertex AI
    aiplatform.init(project=project_id, location=location)
    
    # Set up training job
    job = aiplatform.CustomTrainingJob(
        display_name=f"molecular_generation_{model_type}",
        script_path="src/train.py",
        container_uri="gcr.io/cloud-aiplatform/training/pytorch-gpu.1-8",
        requirements=["src/requirements.txt"],
    )
    
    # Define training parameters
    training_args = {
        "project": project_id,
        "location": location,
        "model_display_name": f"molecular_generation_{model_type}",
        "dataset_path": dataset_path,
        "model_type": model_type,
        "hyperparameters": hyperparameters or {},
    }
    
    # Run training job
    job.run(
        args=training_args,
        replica_count=1,
        machine_type="n1-standard-4",
        accelerator_type="NVIDIA_TESLA_T4",
        accelerator_count=1,
        job_display_name=f"molecular_generation_{model_type}_training",
    )

if __name__ == "__main__":
    # Get environment variables
    project_id = os.getenv("PROJECT_ID")
    location = os.getenv("LOCATION", "us-central1")
    dataset_path = os.getenv("DATASET_PATH")
    model_type = os.getenv("MODEL_TYPE", "vae")
    
    if not all([project_id, dataset_path]):
        raise ValueError("Missing required environment variables")
    
    # Define default hyperparameters
    default_hyperparameters = {
        "learning_rate": 0.001,
        "batch_size": 32,
        "epochs": 100,
        "latent_dim": 128,
    }
    
    # Train model
    train_model(
        project_id=project_id,
        location=location,
        dataset_path=dataset_path,
        model_type=model_type,
        hyperparameters=default_hyperparameters,
    ) 