molecular-ai/requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
rdkit>=2023.3.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
matplotlib>=3.7.0
seaborn==0.12.0
google-cloud-aiplatform>=1.18.0
tqdm>=4.62.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
python-dotenv>=0.19.0
pytest>=6.2.5
black>=21.7b0
flake8>=3.9.2
mypy>=0.910
shap>=0.40.0
lime>=0.2.0.1
deepchem==2.7.1
captum>=0.7.0 